Conflicts:
	plugins/pluginsreaderproviders/iso7816/commands/desfireiso7816commands.cpp

+++ scripts/cleancmake.bat
@@ -1,2 +1,5 @@
+if [%1]==[] goto del
+cd samples/basic
+:del
 del CMakeCache.txt /Q /S
 for /d /r . %%d in (CMakeFiles) do @if exist "%%d" rd /s/q "%%d"
